refactor: change types folder root
