+++ src/pages/AuthPage/Auth/AuthForm/AuthForm.tsx
@@ -5,7 +5,7 @@ import Input from '@/components/Input/Input';
 import AuthButton from '../AuthButton/AuthButton';
 
 import { StyledInputWrapper } from './AuthForm.style';
-import { AuthFormProps } from '@/types/auth';
+import { AuthFormProps } from '@/utils/types/auth';
 
 export default function AuthForm(
   {
