+++ src/utils/types/avatar.ts
